Switch to double precision, added a makefile and other files required to build the fortran functions, and added an algorithm.rg that runs the code. Currently has a variety of debugging asserts as NaNs appear after 14 timesteps

+++ ocean/nemo/nemolite2d/manual_versions/regent/kind_params_mod.f90
@@ -6,7 +6,7 @@ module kind_params_mod
   public
 
   !> Douple precision kind parameter
-  integer, parameter :: GO_WP = c_float !SELECTED_REAL_KIND(12,307)
+  integer, parameter :: GO_WP = c_double !SELECTED_REAL_KIND(12,307)
 
   ! Kind type for double precision
   integer, parameter :: GO_DP = c_double
